Organize things into smaller more focused modules
Make `money_in` and `money_out` into folders that contain modules
relevant for those flows. This especially decomposes money_in into
smaller modules.
